# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Country:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.MyMaster:
    properties:
      clientID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      masterID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a beauty services platform API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BeautyTON API
  version: "1.0"
paths:
...
  /countries:
    post:
      consumes:
      - application/json
      description: Create a new country with the input payload
      parameters:
      - description: Create country
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/entity.Country'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Country'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new country
      tags:
      - countries
  /countries/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a country by country ID
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a country
      tags:
      - countries
    get:
      consumes:
      - application/json
      description: Get country details by country ID
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Country'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a country by ID
      tags:
      - countries
    put:
      consumes:
      - application/json
      description: Update country details by country ID
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: string
      - description: Update country
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/entity.Country'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Country'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a country
      tags:
      - countries
  /files:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to the server
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a file
      tags:
      - files
  /files/{id}:
    get:
      consumes:
      - application/json
      description: Download file by file ID
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a file by ID
      tags:
      - files
...
  /my_masters:
    post:
      consumes:
      - application/json
      description: Create a new my master with the input payload
      parameters:
      - description: Create my master
        in: body
        name: mymaster
        required: true
        schema:
          $ref: '#/definitions/entity.MyMaster'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.MyMaster'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new my master
      tags:
      - my-masters
  /my_masters/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a my master by my master ID
      parameters:
      - description: My Master ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a my master
      tags:
      - my-masters
    get:
      consumes:
      - application/json
      description: Get my master details by my master ID
      parameters:
      - description: My Master ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.MyMaster'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a my master by ID
      tags:
      - my-masters
    put:
      consumes:
      - application/json
      description: Update my master details by my master ID
      parameters:
      - description: My Master ID
        in: path
        name: id
        required: true
        type: string
      - description: Update my master
        in: body
        name: mymaster
        required: true
        schema:
          $ref: '#/definitions/entity.MyMaster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.MyMaster'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a my master
      tags:
      - my-masters
...
swagger: "2.0"